add RetinaNetModel and DeepForestModule to models.py
